♻️ (#391) 프로젝트 떠나기 모달 리팩터링
- ProjectLeaveModal 폴더로 이동
- 리팩터링한 훅 적용 (UI 관련 처리 외부 위임 방식 적용)
- 그 외 코드 및 클래스 정리
